avoid of the remaining warnings during WB build on windows
Review by Mike [id: 8722]

+++ modules/wb.fabric.interface/wb_fabric_interface.cpp
@@ -50,7 +50,7 @@ int WbFabricInterfaceImpl::openConnection(const db_mgmt_ConnectionRef &conn, con
 
   // Sets the connection timeout
   grt::DictRef wb_options = grt::DictRef::cast_from(get_grt()->get("/wb/options/options"));
-  int connect_timeout = wb_options.get_int("Fabric:ConnectionTimeOut", 60);
+  unsigned int connect_timeout = (unsigned int)wb_options.get_int("Fabric:ConnectionTimeOut", 60);
   mysql_options(&mysql, MYSQL_OPT_CONNECT_TIMEOUT, &connect_timeout);
